+++ src/utils/setupOutputFileSystem.js
@@ -1,42 +1,36 @@
-import path from "path";
+const memfs = require("memfs");
 
-import { createFsFromVolume, Volume } from "memfs";
+/** @typedef {import("webpack").MultiCompiler} MultiCompiler */
+/** @typedef {import("../index.js").IncomingMessage} IncomingMessage */
+/** @typedef {import("../index.js").ServerResponse} ServerResponse */
 
-export default function setupOutputFileSystem(context) {
+/**
+ * @template {IncomingMessage} Request
+ * @template {ServerResponse} Response
+ * @param {import("../index.js").Context<Request, Response>} context
+ */
+function setupOutputFileSystem(context) {
   let outputFileSystem;
 
   if (context.options.outputFileSystem) {
-    // eslint-disable-next-line no-shadow
     const { outputFileSystem: outputFileSystemFromOptions } = context.options;
-
-    // Todo remove when we drop webpack@4 support
-    if (typeof outputFileSystemFromOptions.join !== "function") {
-      throw new Error(
-        "Invalid options: options.outputFileSystem.join() method is expected"
-      );
-    }
-
-    // Todo remove when we drop webpack@4 support
-    if (typeof outputFileSystemFromOptions.mkdirp !== "function") {
-      throw new Error(
-        "Invalid options: options.outputFileSystem.mkdirp() method is expected"
-      );
-    }
 
     outputFileSystem = outputFileSystemFromOptions;
   } else {
-    outputFileSystem = createFsFromVolume(new Volume());
-    // TODO: remove when we drop webpack@4 support
-    outputFileSystem.join = path.join.bind(path);
+    outputFileSystem = memfs.createFsFromVolume(new memfs.Volume());
   }
 
-  const compilers = context.compiler.compilers || [context.compiler];
+  const compilers =
+    /** @type {MultiCompiler} */
+    (context.compiler).compilers || [context.compiler];
 
   for (const compiler of compilers) {
-    // eslint-disable-next-line no-param-reassign
     compiler.outputFileSystem = outputFileSystem;
   }
 
+  // @ts-ignore
   // eslint-disable-next-line no-param-reassign
   context.outputFileSystem = outputFileSystem;
 }
+
+module.exports = setupOutputFileSystem;